+++ lib/NfcoreSchema.groovy
@@ -70,7 +70,6 @@ class NfcoreSchema {
             'quiet',
             'syslog',
             'v',
-            'version',
 
             // Options for `nextflow run` command
             'ansi',