+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [next]
+### Added
+* `nextflow run epi2me-labs/wf-human-variation --version` will now print the workflow version number and exit
 ### Changed
 * `--modbam2bed_args` can be used to further configure the wf-methylation `modbam2bed` process
 * `modbam2bed` outputs are now prefixed with `<sample_name>.methyl`
+### Removed
+* `-profile conda` is no longer supported, users should use `-profile standard` (Docker) or `-profile singularity` instead
 
 ## [v0.4.1]
 ### Fixed